+++ bin/deploy.js
@@ -6,6 +6,7 @@ var exec = require('child_process').exec,
     rl = require('readline'),
     conf = require('../config');
 
+var deployMessage = process.argv[2];
 process.chdir(conf.deployDir);
 
 //create CNAME file
@@ -21,7 +22,7 @@ exec('cp -R '+ conf.publicDir +'/* '+ conf.deployDir, function(code, stdout, std
         exec('git add -u', function(code, stdout, stderr) {
             console.log(stdout.white);
             console.log('haroo> git commit -m ¶'.yellow);
-            var message = 'Site updated at '+ new Date();
+            var message = deployMessage || 'Site updated at '+ new Date();
             exec('git commit -m "'+ message +'"', function(code, stdout, stderr) {
                 console.log(stdout.white);
                 console.log('haroo> git push origin master --force¶'.yellow);