+++ config.js
@@ -5,6 +5,7 @@ module.exports = {
         "description": "하루프레스는 하루를 손쉽게 기록해 보자라는 뜻의 블로그 엔진으로, Node.js 와 EJS 템플릿 엔진으로 설계된 Github 에서 제공하는 정적 페이지 서비스를 위한 블로그 엔진입니다.",
         "siteUrl": "http://haroopress.github.com",
         "author": "Rhio Kim",
+        "email": "rhio.kim@gmail.com",
         "keywords": [
             "node.js",
             "haroopress",
@@ -18,8 +19,9 @@ module.exports = {
         ]
     },
     "lang": "en",
-    "contentLength": 3,
+    "contentLength": 5,
     "pagenate": 5,
+    "dateFormat": "MMM Do 'YY",
     "deployBranch": "gh-pages",
     "CNAME": "",
     "sourceDir": __dirname +"/source/data",
@@ -46,8 +48,8 @@ module.exports = {
             "skipForks": true
         },
         "tweets": {
-            "display": false,
-            "user": "",
+            "display": true,
+            "user": "haroopress",
             "tweetCount": 10
         },
         "twitter": {
@@ -71,7 +73,9 @@ module.exports = {
         },
         "delicious": {},
         "contributors": {
-            "display": true
+            "display": true,
+            "sort": 'DESC',
+            "count": 5
         },
         "weather": {
             "display": true,